# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export type ScheduleTranslationRequestTargetLanguage = typeof ScheduleTranslationRequestTargetLanguage[keyof typeof ScheduleTranslationRequestTargetLanguage];


export const ScheduleTranslationRequestTargetLanguage = {
  hi: 'hi',
  mr: 'mr',
} as const;

export interface ScheduleTranslationItem {
  /** @maxLength 100 */
  id: string;
  /** @maxLength 300 */
  title: string;
  /** @maxLength 2500 */
  description: string;
  /** @maxLength 300 */
  location: string;
}

export interface ScheduleTranslationRequest {
  targetLanguage: ScheduleTranslationRequestTargetLanguage;
  /**
     * @minItems 1
     * @maxItems 40
     */
  items: ScheduleTranslationItem[];
}

export type ScheduleTranslationResponseTargetLanguage = typeof ScheduleTranslationResponseTargetLanguage[keyof typeof ScheduleTranslationResponseTargetLanguage];


export const ScheduleTranslationResponseTargetLanguage = {
  hi: 'hi',
  mr: 'mr',
} as const;

export interface TranslatedScheduleItem {
  id: string;
  title: string;
  description: string;
  location: string;
}

export interface ScheduleTranslationResponse {
  targetLanguage: ScheduleTranslationResponseTargetLanguage;
  items: TranslatedScheduleItem[];
}