# lib/api-zod/src/generated/types/scheduleTranslationItem.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */

export interface ScheduleTranslationItem {
  /** @maxLength 100 */
  id: string;
  /** @maxLength 300 */
  title: string;
  /** @maxLength 2500 */
  description: string;
  /** @maxLength 300 */
  location: string;
}