# lib/api-zod/src/generated/types/scheduleTranslationRequest.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */
import type { ScheduleTranslationItem } from './scheduleTranslationItem';
import type { ScheduleTranslationRequestTargetLanguage } from './scheduleTranslationRequestTargetLanguage';

export interface ScheduleTranslationRequest {
  targetLanguage: ScheduleTranslationRequestTargetLanguage;
  /**
     * @minItems 1
     * @maxItems 40
     */
  items: ScheduleTranslationItem[];
}