# lib/api-zod/src/generated/types/scheduleTranslationResponse.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */
import type { ScheduleTranslationResponseTargetLanguage } from './scheduleTranslationResponseTargetLanguage';
import type { TranslatedScheduleItem } from './translatedScheduleItem';

export interface ScheduleTranslationResponse {
  targetLanguage: ScheduleTranslationResponseTargetLanguage;
  items: TranslatedScheduleItem[];
}