# lib/api-zod/src/generated/types/translatedScheduleItem.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */

export interface TranslatedScheduleItem {
  id: string;
  title: string;
  description: string;
  location: string;
}